requirements.txt
wsgiref==0.1.2
gunicorn==19.0.0
Django
django-chunked-upload
psycopg2
django-cors-headers
django-jwt-auth
djangorestframework-jwt
djangorestframework
